Optimize '_' to ' ' replacement in MediaWiki links with Qt5
The speed-up is achieved by traversing the article text once instead of
iterating many times over it.

The underscore to space replacement code works 10-50 times faster with
this change on tested articles ("United States", "Paris" from English
Wikipedia and "Anakin Skywalker/Legends" from English Wookieepedia).

I tried to implement this optimization both for Qt4 and Qt5. However
similar code using QRegExp turned out to be up to 20 times slower than
the original code on the tested articles. This dramatic slowdown is
caused by the fact that QRegExp::indexIn() copies and stores the subject
string str internally, so changing the string by replacing '_' with ' '
in the link detaches it and makes a deep copy. This introduces massive
overhead for long articles. It would probably be faster to store
start and end positions of each link in a container; then, after
destroying the QRegExp object, replace '_' with ' ' in each remembered
string segment; but this solution is going to be much less elegant
and slower than the straightforward QRegularExpression implementation.

Testing platform: Qt 5.10.0, GCC 7.2.1 on Linux 64-bit.