Add an option to avoid auto-scrolling out of top dictionary
When a user clicks a link in a dictionary or requests translation of
a word by double-clicking or translates selection via the context menu,
the definition in all dictionaries is displayed. Then the article from
the dictionary, out of which the translation was requested, becomes
current and the view scrolls down to this article placing it on top,
hiding translations from dictionaries above it.

This behavior is inconvenient in some workflows and forces the user to
manually navigate to the top dictionary translation each time this
auto-scrolling happens.

For example: a user has English->Russian dictionaries and
English->English dictionaries. The English->Russian dictionaries are
higher up in the dictionary order because they provide easier/faster to
understand translations. Some rare words and phrases are missing from
the English->Russian dictionaries however. So the user has to read the
English explanation of the word/phrase. When the user double-clicks a
word or follows a link in the English->English dictionary article, she
would rather see translations from the preferable English->Russian
dictionaries.

The new option allows to disable auto-scrolling and ensure that articles
from higher-priority dictionaries are visible. The option doesn't affect
backward/forward navigation via arrow buttons or Alt+Arrow shortcuts:
these still scroll to the stored vertical position among articles. This
remaining auto-scrolling is not a problem for the described use case and
hopefully for other potential use cases.