Search input phrase with punctuation too
This is a prototype implementation. A proper implementation should pass
a struct/class (that contains the word string and the two indices)
through many GoldenDict's interfaces.

This implementation seems to work well with DSL dictionaries, but not so
well with MediaWiki (see the 3rd bug below).

An advantage of defining PRIMARY_WORD_NO_PUNCT as 1 is that when there
is no translation for the word, GoldenDict informs: "No translation for
<the input phrase with trimmed punctuation> was found in group <X>".
Had it displayed say a comma at the end of the not-found word, a user
might think that no translation has been found because of the comma.

When for example "etc." is sent for translation and a dictionary
contains translations for both "etc." and "ETC", the value of
PRIMARY_WORD_NO_PUNCT determines which article is displayed above.

Known bugs:
1. A word with '\0' symbols and indices is stored in History. This bug
will disappear in the proper implementation.
2. A word with untrimmed punctuation is displayed in the scan popup's
search line even if PRIMARY_WORD_NO_PUNCT equals 1. This can be
rectified in the prototype implementation by writing more code at the
site of the small change this commit makes in scanpopup.cc. This bug
will disappear in the proper implementation.
3. Select "e.g." (w/o quotes), Ctrl+C+C => English Wiktionary shows two
identical articles one after another. I don't know how difficult it will
be to fix this issue.